# Performance improvements for IdDict