--- a/base/iddict.jl
+++ b/base/iddict.jl
@@ -29,7 +29,8 @@ mutable struct IdDict{K,V} <: AbstractDict{K,V}
     ht::Memory{Any}
     count::Int
     ndel::Int
-    IdDict{K,V}() where {K, V} = new{K,V}(Memory{Any}(undef, 32), 0, 0)
+    age::UInt
+    IdDict{K,V}() where {K, V} = new{K,V}(Memory{Any}(undef, 32), 0, 0, 0)
 
     function IdDict{K,V}(itr) where {K, V}
         d = IdDict{K,V}()
@@ -44,7 +45,7 @@ mutable struct IdDict{K,V} <: AbstractDict{K,V}
         d
     end
 
-    IdDict{K,V}(d::IdDict{K,V}) where {K, V} = new{K,V}(copy(d.ht), d.count, d.ndel)
+    IdDict{K,V}(d::IdDict{K,V}) where {K, V} = new{K,V}(copy(d.ht), d.count, d.ndel, d.age)
 end
 
 IdDict() = IdDict{Any,Any}()
@@ -60,7 +61,10 @@ IdDict(kv) = dict_with_eltype((K, V) -> IdDict{K, V}, kv, eltype(kv))
 empty(d::IdDict, ::Type{K}, ::Type{V}) where {K, V} = IdDict{K,V}()
 
 function rehash!(d::IdDict, newsz = length(d.ht)%UInt)
+    age0 = d.age
     d.ht = ccall(:jl_idtable_rehash, Memory{Any}, (Any, Csize_t), d.ht, newsz)
+    @assert d.age == age0 "Multiple concurrent writes to Dict detected!"
+    d.age += 1
     d
 end
 
@@ -85,6 +89,7 @@ end
 @propagate_inbounds function _setindex!(d::IdDict{K,V}, val::V, keyindex::Int) where {K, V}
     d.ht[keyindex+1] = val
     d.count += 1
+    d.age += 1
 
     if d.ndel >= ((3*length(d.ht))>>2)
         rehash!(d, max((length(d.ht)%UInt)>>1, 32))
@@ -99,6 +104,7 @@ function setindex!(d::IdDict{K,V}, @nospecialize(val), @nospecialize(key)) where
     end
     keyindex = ht_keyindex2!(d, key)
     if keyindex >= 0
+        d.age += 1
         @inbounds d.ht[keyindex+1] = val
     else
         @inbounds _setindex!(d, val, -keyindex)
@@ -120,6 +126,7 @@ end
 function pop!(d::IdDict{K,V}, @nospecialize(key), @nospecialize(default)) where {K, V}
     found = RefValue{Cint}(0)
     val = ccall(:jl_eqtable_pop, Any, (Any, Any, Any, Ptr{Cint}), d.ht, key, default, found)
+    d.age += 1
     if found[] === Cint(0)
         return default
     else
@@ -148,6 +155,7 @@ function empty!(d::IdDict)
     @_gc_preserve_end t
     d.ndel = 0
     d.count = 0
+    d.age += 1
     return d
 end
 
@@ -171,11 +179,17 @@ function get!(d::IdDict{K,V}, @nospecialize(key), @nospecialize(default)) where 
     if keyindex < 0
         # If convert call fails we need the key to be deleted
         d.ndel += 1
+        age0 = d.age
         val = isa(default, V) ? default : convert(V, default)::V
         d.ndel -= 1
-        @inbounds _setindex!(d, val, -keyindex)
+        if d.age != age0
+            @inline setindex!(d, val, key)
+        else
+            @inbounds _setindex!(d, val, -keyindex)
+        end
         return val::V
     else
+        d.age += 1
         return @inbounds d.ht[keyindex+1]::V
     end
 end
@@ -190,16 +204,26 @@ function get(default::Callable, d::IdDict{K,V}, @nospecialize(key)) where {K, V}
 end
 
 function get!(default::Callable, d::IdDict{K,V}, @nospecialize(key)) where {K, V}
-    val = ccall(:jl_eqtable_get, Any, (Any, Any, Any), d.ht, key, secret_table_token)
-    if val === secret_table_token
+    keyindex = ht_keyindex2!(d, key)
+
+    if keyindex < 0
+        # If convert call fails we need the key to be deleted
+        d.ndel += 1
+        age0 = d.age
         val = default()
         if !isa(val, V)
             val = convert(V, val)::V
         end
-        setindex!(d, val, key)
-        return val
-    else
+        d.ndel -= 1
+        if d.age != age0
+            @inline setindex!(d, val, key)
+        else
+            @inbounds _setindex!(d, val, -keyindex)
+        end
         return val::V
+    else
+        d.age += 1
+        return @inbounds d.ht[keyindex+1]::V
     end
 end