--- a/test/dict.jl
+++ b/test/dict.jl
@@ -1639,3 +1639,13 @@ end
         @test                    valtype(D{K,    V}) == V
     end
 end
+
+# Check mutating IdDict during get! works
+let d = IdDict()
+    function f()
+        d[1] = 4
+        return -2
+    end
+    @test get!(f, d, 1) === -2
+    @test length(d) == 1
+end