The main improvement is from `get!` which no longer calculates the hash twice when it has to setindex and so is twice as fast.

This is useful for a global value numbering pass I've written.