@JeffBezanson 2023-08-29
Thanks, any speedup numbers to whet the appetite?

I'm not really comfortable duplicating all the lookup code, and the try-catch around the conversion is not a good pattern to use. This needs a different design. Maybe we can copy how `get!` works for `Dict`.