@Zentrik 2023-08-30
Here's some performance numbers,
```jl
function f!(d)
    for x in 1:10^4
        d[x] = 1
    end
end

function g!(d)
   for x in 1:10^4
       get!(d, x, 1)
   end
end
```
On master
```jl
julia> versioninfo()
Julia Version 1.11.0-DEV.356
Commit 4768d7ab8fe (2023-08-28 12:54 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 16 × AMD Ryzen 7 1700 Eight-Core Processor
  WORD_SIZE: 64
  LLVM: libLLVM-15.0.7 (ORCJIT, znver1)
  Threads: 1 on 16 virtual cores

julia> @btime f!(d) setup=(d=IdDict{Int, Int}(); sizehint!(d, 10^4))
  409.015 μs (9489 allocations: 148.27 KiB)

julia> @btime g!(d) setup=(d=IdDict{Int, Int}(); sizehint!(d, 10^4))
  639.763 μs (18978 allocations: 296.53 KiB)
```
On this pr
```jl
@btime f!(d) setup=(d=IdDict{Int, Int}(); sizehint!(d, 10^4))
  324.515 μs (9489 allocations: 148.27 KiB)

@btime g!(d) setup=(d=IdDict{Int, Int}(); sizehint!(d, 10^4))
  330.156 μs (9489 allocations: 148.27 KiB)
```